ODC - Becoming a Web Developer - Building  Web Form Exercise
Passed validation. 288/288
The validator is wrong. It wants the Job Title label in the Employee Details form to be "Job Title Id" instead. SImilarly in the Project Member form, it wants the Employee label and the Project label to say "Employee Id" and "Project Id" respectively. Maybe it's from an old version of ODC, but the current version correctly does not add "Id" to the labels. I will correct it in the next check-in.