ODC - Becoming a Web Developer - Fetching Data On Demand Exercise
No code validation.
Added Country static entity, and State and City entities. Added CityId to Employee.
Added dropdowns to EmployeeDetails to select country, state, and city.
Changed aggregates to fire on demand and to call dropdown on change handlers on after fetch.
Data was added to the database that needs to be moved to a Excel bootstrap.
There's a persistent bug in the picture upload.